requirements.txt
pandas
numpy
vectorbt
yfinance
ccxt
ib_insync
matplotlib
ta
pyyaml
loguru
websockets
python-dotenv
fastapi
uvicorn
anywidget
mplfinance
